Remove monitoring UDP hub_port from example/test configs
This port will be chosen dynamically, and that is fine in test situations.

Ongoing work to refactor monitoring radios will remove this hub_port
parameter, making it either unnecessary (when non-UDP monitoring radios are
used) or specified as part of the radio selection.

This PR removes hub_port from tests and examples as preparation for that
change.