Remove monitoring UDP hub_port from example/test configs (#3710)
This port will be chosen dynamically, and that is fine in test
situations.

This is probably also better in environments where users run multiple
Parsl instances at once: they cannot all use the removed port 55055 and
will interfere with each other.

Ongoing work to refactor monitoring radios will remove this hub_port
parameter, making it either unnecessary (when non-UDP monitoring radios
are used) or specified as part of the radio selection.

This PR removes hub_port from tests and examples as preparation for that
change.

# Changed Behaviour

The monitoring UDP port will be chosen dynamically in tests and
examples.

## Type of change

- Code maintenance/cleanup