Move execute_wait from vestigial LocalChannel into parsl.utils (#3705)
# Description

This moves the execute_wait functionality previously provided by
LocalChannel into parsl.utils, as part of #3515.
LocalChannel.execute_wait did not reference `self` so it already
basically behaved as a function rather than a method.

This leaves LocalChannel as solely a place for script_directory, which
will be untangled in a subsequent PR.

# Changed Behaviour

none

## Type of change

- Code maintenance/cleanup