Remove channel classes and parameters (#3716)
Before this PR, all channel functionality was removed and the channel
classes and config parameters remained as stubs waiting to be removed.
This PR removes them.

# Changed Behaviour

Any place that a user is specifying `channel=LocalChannel()` in their
configuration will now break. That option can be removed by the user
because that option should not have any effect in the codebase
immediately prior to this PR.

## Type of change

- Code maintenance/cleanup