Fix rounding error in htex block scale in (#3721)
# Description

PR #2196 calculates a number of blocks to scale in, in the htex
strategy, rather than scaling in one block per strategy iteration.
However, it rounds the wrong way: it scales in a rounded up, rather than
rounded down, number of blocks.

Issue #3696 shows that then resulting in oscillating behaviour: With 14
tasks and 48 workers per block, on alternating strategy runs, the code
will either scale up to the rounded up number of needed blocks (14/48 =>
1), or scale down to the rounded down number of needed blocks (14/48 =>
0).

This PR changes the rounding introduced in #2196 to be consistent:
rounding up the number of blocks to scale up, and rounding down the
number of blocks to scale down.

# Changed Behaviour

HTEX scale down should oscillate less

# Fixes

Fixes #3696 

## Type of change

- Bug fix