Upgrade SQLAlchemy to version 2 (#3734)
This is part of supporting Python 3.13 (PR #3646), as a prereq to
upgrading pandas.

# Changed Behaviour

Development environments (which have the requirements in
`test-requirements.txt` installed) will break and need `pip uninstall
sqlalchemy2-stubs` - sqlalchemy packaging does not express a
negative/removal requirement against that transitional package.

non-parsl user code which happens to use the parsl-installed sqlalchemy
might break - for example, see API changes in the test suite in this PR

## Type of change

- Code maintenance/cleanup