Tidy DataFuture argument types
The task ID is never optional: a DataFuture always comes from a
Parsl task. (in the same way that an AppFuture always has a
task record)

Long ago, Files could be specified as strings. This behaviour
was removed slightly less long ago. Right now, typeguard should
be detecting if a string is passed, because of the type signature,
and so the explicit type check for strings should be unnecessary.
This PR removes that explicit type check.

The future being tracked by the data future is not necessarily an
AppFuture: it can be any kind of future that completes to indicate
the represented file is now available. That is already reflected
in the type annotations and code, but this PR fixes the docstring.