Change some exceptions to put human readable text in __str__ not __repr__ (#3741)
# Description

See #2025 for more context.

# Changed Behaviour

Rendering of exceptions made with `__repr__` will change to a more
machine readable format. Generally Python itself reports exceptions
using `__str__`, but DFK task failure history (which goes into the
monitoring database) is rendered with `__repr__`. This is probably
better behaviour for that history field, which is
quasi-machine-readable.

## Type of change

- Update to human readable text: Documentation/error messages/comments
- Code maintenance/cleanup