Test interchange exit on bad registration message (#3698)
On certain bad registration messages, the interchange should exit
immediately. This tests that.

See #3697 for some bad (cosmetic?) behaviour here - the interchange
SIGABRTs on this code path rather than exiting cleanly, and this test
includes a commented out assert that could check for clean exit (in
addition to checking that the interchange process exits at all)

## Type of change

- Code maintenance/cleanup