Revert part of PR #3638 which is part of a now-abandoned task priority PR series (#3750)
PR #3638 introduced parameters and documentation for a task priority
field which was never implemented in mainline Parsl. This PR leaves most
of the infrastructure in place, as other code now depends on it, but
removes `priority` as a valid resource specification field.

# Changed Behaviour

users trying to use the no-effect priority parameter will now get
runtime errors rather than silent ignore.

## Type of change

- Code maintenance/cleanup