Remove quasi-monitoring checkpoint file dfk.pkl (#3760)
This file is saved alongside checkpoints, but stores an extremely light
summary of DFK state rather than of checkpoints: the DFK's task count
and run directory path.

That is information that falls into the space of monitoring, not of
checkpointing - so this PR removes that file, in favour of the much
richer monitoring database.


# Changed Behaviour

Checkpoints will no longer contain the light monitoring data that was
previously in `dfk.pkl`. Users will have to get that information some
other way (for example, by looking at the location of the checkpoint
directory to discover the run directory, and by using Parsl's monitoring
facility for task counts)

## Type of change

- Code maintenance/cleanup