Add command to get installed packages on managers (#3773)
# Description

Added a `MANAGERS_PACKAGES` HTEX interchange command and associated HTEX
method to retrieve a dict mapping each manager ID to a dict of installed
packages and their versions. This is useful for debugging and monitoring
worker environments.

We gather the package information from the manager upon registration,
but do not include it in the standard `MANAGERS` command, which runs
every 5 seconds. This avoids unnecessary overhead.

## Type of change

- New feature