Refactor result loop, remove timeout
There is no need to iterate the loop to send a heartbeat.  Instead, use a
dedicated thread (heartbeater), and no longer send messages to ZMQ in bulk.
ZMQ already handles messages messages performantly and `_multipart()` is for an
_atomic_ send of multiple messages or message parts.

This reduces the complexity of the loop, and also reduces the idle CPU load.
On my local hardware, it reduces the idle load by about 40% (not scientific;
just looking at htop).

In future work, can replace `result_outgoing.send_multipart()` with just
`result_outgoing.send()`