Create and close ZMQ sockets on utilizing thread (#3774)
The ZMQ documentation is quite clear about not sharing ZMQ sockets
between threads. This case wasn't so much shared as simply created on
the main thread and _used_ in the child threads. "Probably" okay, and
hasn't appeared to have bitten us to date ... regardless, be cleaner,
and in so doing recognize that these aren't actually class variables.

# Changed Behaviour

Slightly different logging output, as the result and task ZMQ pipes are
now connecting in different threads.

## Type of change

- Code maintenance/cleanup