Run slurm and flux tests in merge queue (#3787)
In order to make more tests mandatory, those tests need to run in the
merge queue. PR #3661 did this for the main CI tests. This PR does it
for flux and slurm tests so that they can be made mandatory (in github
non-versioned config)

# Changed Behaviour

more tests will run at merge time. nothing user-facing.

## Type of change

- Code maintenance/cleanup