Make interchange monitoring ZMQ use loopback, like other connections (#3795)
The HighThroughputExecutor assumes that the submit process and the
interchange are in the same host (in the sense that the configured
loopback address will work to connect from one to the other).

This PR adds that assumption for connecting from the interchange to the
monitoring router. Which is pretty much already a constraint: that
router is launched from the submit process using Python multiprocessing
and so will run on the same host as the submit process.

# Changed Behaviour

in-host network connections for monitoring will look a bit different,
but shouldn't be a behaviour change

## Type of change

- Code maintenance/cleanup