Split MonitoringRouter into ZMQ and UDP processes (#3796)
This is to facilitate further work where the ZMQ and UDP router
processes can have different lifetimes (along with the already separated
filesystem router process and any future plugins).

The ZMQ router process is only needed when HTEX is being used, to route
messages from the interchange.

The UDP router process is only needed when resource monitoring is
enabled, and an executor is configured with the udp radio mode.

This PR does not change when these two processes run, but is only
intended to help future PRs make this change.

# Changed Behaviour

one more process at runtime when monitoring is enabled

## Type of change

- Code maintenance/cleanup