Fix bug in MPITaskScheduler where tasks with identical priorities fail with `TypeError` (#3794)
# Description

The `MPITaskScheduler` uses Python's PriorityQueue to prioritize tasks
based on the number of nodes requested.
When items with identical priorities are are submitted to the
PriorityQueue, they attempt to sort based on the task dict which fails
with TypeError unhashable type: dict. This PR adds a new
`PrioritizedTask` dataclass that sets the task element to
`field(compare=False)`.

I'm splitting changes in #3783 to keep the PR concise. This is split 1
of 3.


# Changed Behaviour

Fixes the bug described above.

## Type of change

Choose which options apply, and delete the ones which do not apply.

- Bug fix