Make monitoring shutdown always event-terminate-kill-close
Prior to this PR, the monitoring proceses were shut down in two ways:

* In the happy path, processes would be notified that it was shutdown
  time via a multiprocessing.Event(). They would be joined on, with
  no timeout, assuming that they would eventually shut down based on
  the Event. In the unhappy case that one does not shut down,
  DFK shutdown would hang here.

* In the unhappy path, processes each have an opportunity to notify
  that an exception has happened via the exception_q. This is used
  for two purposes: to transcribe the exception into the main Parsl
  log, and to change how shutdown happens: if any exception has been
  delivered this way by any of the monitoring processes, then shutdown
  will happen by calling .terminate() on each process, without event
  notification.

This PR changes that so the exception_q is ignored for termination
purposes (so it becomes purely a remote-logging system).

In both situations, processes will be ended using a sequence:

* notify by event

* join with timeout

* .terminate()

* join with timeout

* .kill()

The timeout for join is set quite high: monitoring processes sometimes
take a while to shut down, and the timeout path is an error path not
the fast happy path.

Changed user-facing behaviour: monitoring processes which take longer
than the timeout but would still otherwise successfully shut down
will now be .terminate()ed, meaning whatever they were usefully doing
will not happen.

Shutdown in the failing case might take longer: if a process reports
an exception and then hangs, it will now take the timeout period to
be terminated, rather than immediately.

If a process hangs without reporting an exception, it will now
eventually be terminated rather than shutdown hanging.

TODO:
perhaps not specifically in the monitoring shutdown, but in some more unit-testy deliberately broken
processes?