Make database manager exit by Event, rather than STOP message (#3807)
This is for consistency with other monitoring process exit.

The database manager already has an internal exit event for shutdown of
threads. This PR does not change handling there: the event is set inside
the close() method which is also called in other situations, so it isn't
obviously the "same" kind of exit event.

For minimality of change, the exit event is checked in the same polling
thread as previously was polling for the STOP message.

There's an expectation that the DBM will stay alive until after the
monitoring radio routers have shut down - this is part of loosely
defined race-y shutdown behaviour around executor level task completion
vs delivery of monitoring messages over radios: without this,
occasionally the DBM will shut down very fast, before all task
completion messages have been put into the database, and tests will
fail.

To preserve that ordering, this new DBM shutdown uses a separate
shutdown event, which is set at the point that the STOP message was
previously sent.

# Changed Behaviour

peturbation in shutdown timings which might now show new race conditions

## Type of change

- Code maintenance/cleanup