Remove exception_q from monitoring (#3812)
After PR #3811 removed exeception_q's role in exit handling, this code
remains only as a rudimentary distributed logging system to convey some
specific error messages back to the main process for logging in
parsl.log.

No other part of Parsl does this in this way, and instead each process
logs to a file with the user expected to look in individual log files.

This PR makes monitoring consistent with that expectation.

# Changed Behaviour

Error messages from monitoring failures will appear in different places,
compared to before this PR.

## Type of change

- Update to human readable text: Documentation/error messages/comments
- Code maintenance/cleanup