Add monitoring info for tasks given to a manager (#3833)
# Description

Update any interested listeners to the state of the manager when new
tasks are assigned. Based on `.start()`, this now covers most (all?)
relevant changes to a manager's info block that will be shared.

# Changed Behaviour

Adds a monitoring message per manager in the interchange loop, for any
outstanding tasks that move from the interchange to a worker manager. I
don't think any existing workflows will be impacted, but there will now
be an extra NODEINFO record in the sqlite database per task.

## Type of change

- New feature