Add result-side monitoring update (#3837)
# Description

This is a followup to #3833; had meant to add it in there after #3834,
but miscommunicated so adding now.

# Changed Behaviour

Adds a result-side monitoring messages for managers that send results.
No workflows should be impacted, but there is now an additional NODEINFO
record in the sqlite database. This PR will bring the total up to 4 (as
shown in the changed test).

## Type of change

- New feature