Move WorkQueue and TaskVine tests to a separate GH Action (#3846)
This PR splits out WorkQueue and TaskVine into a separate Github action so that optional extra packages can be tested separately. This allows us to 1) identify test failures originating from core parsl vs external packages and 2) test different components in parallel. 

The current mechanism that relies on pulling binaries for cctools is
broken in Ubuntu 24.04, and based on the discussion here:
https://github.com/Parsl/parsl/issues/3544, the recommended solution is
to switch to using binaries distributed through conda.