Make `MPITaskScheduler` prioritize large tasks (#3805)
# Description

The python `PriorityQueue` prioritizes items with smaller priority
values, which leads us to our current `MPITaskScheduler` prioritizing
MPI tasks with fewer nodes requested since the nodes requested is used
as the priority. This PR changes the ordering to prefer larger jobs.

# Changed Behaviour

MPI Tasks with larger node requests are prioritized by the manager.
Please note that as @WardLT pointed out in the comments on #3783, the
greedy scheduling that we use can lead to larger tasks getting delayed
in favor of smaller tasks that can get scheduled immediately.

These are change split from https://github.com/Parsl/parsl/pull/3783 to
keep the PR concise.
This is split 3 of 3.

## Type of change

Choose which options apply, and delete the ones which do not apply.

- Update to human readable text: Documentation/error messages/comments