Fix binding when user sets HTEx worker ports manually (#3838)
# Description

Fixes the creation of the TCP url when the ports used by HTEx workers
are fixed.

The `tcp_url` helper function from addresses.py was only appending the
port number when the address was not "*".

# Changed Behaviour

HTEx works with the ports fixed. Not sure when it was broken

I added a unit test which shows that the current code fails when
launching an HTEx interchange with user-set ports.

# Fixes

None (yet? should I make one?)

## Type of change

- Bug fix