Reduce requirements in parsl_resource_specification test (#3864)
These resources are not actually required for the test, which is only
validating behaviour when the resources are specified or not.

On resource constrained systems, prior to this PR, I often encountered a
test hang as Task Vine's behaviour on insufficent resource is to wait
for the resource to become available.

# Changed Behaviour

tests should be less hangy in resource constrained situations

## Type of change

- Code maintenance/cleanup