Use itertools batch to get long jobs lists (#3815)
# Description

When there are many jobs being tracked by parsl `sacct` can become
unresponsive and timeout if there are too many jobs. A users workflow
was experiencing this while running through Globus Compute on
Perlmutter.

# Changed Behaviour

This should batch calls in groups to Slurm

# Fixes

Fixes #3814

## Type of change

Choose which options apply, and delete the ones which do not apply.

- Bug fix

---------

Co-authored-by: Kevin Hunter Kesling <kevin@globus.org>