Add Add priority queue to HTEX interchange v4 (#3848)
# Description

Replaces pending_task_queue's queue.Queue() as a SortedList which
functions as a priority queue in the interchange. Used to prioritize
certain user-labeled tasks when distributing tasks to managers

# Changed Behaviour

Users can now use the parsl_resource_specification keyword in python
apps to specify a 'priority'. Lower value is higher priority, tasks with
higher priority will be distributed to workers first. If no priority is
given, tasks default to the lowest priority and the FIFO queue behavior
of the original queue.Queue() is retained.

# Fixes

Part of #3323, reverts changes in #3750 

## Type of change

Choose which options apply, and delete the ones which do not apply.

- New feature