Move try_assert to exponential backoff (#3868)
# Description

In networking contexts, the general advice for polling for specific
events is "exponential backoff." I don't think this is related to our
recent CI test failures as 20ms is a super long time to wait (the
standard game frame time is 1/60s or ~16.7ms, for example, and
considered a low-bar to beat) and all of the test infrastructure is on
the same host. Regardless, exponential backoff is good hygiene in
general, so "let's see what happens."

# Changed Behaviour

No user-facing change.

## Type of change

- Code maintenance/cleanup