Remove unused hub address parameter for UDP router (#3877)
This stayed wired in when the ZMQ and UDP routers were split in PR
#3796.

The UDP router does not use the hub address. It is only used by the ZMQ
router.

As an alternative to remove this unused parameter, the UDP router could
be modified to listen only on the supplied address - that remains as an
option in future, but my work right now is about tidy up, not about
adding new features.

## Type of change

- Code maintenance/cleanup