Make try_assert exponential backoff smoother
Before this PR, a sequence of small random number choices could make
doublings happen quite fast, pushing the delay time to be quite high
faster than exponential rate and then causing later iterations to
wait for longer than might be expected.

This PR makes the delay time increase according to the chosen delay
for each iteration.

See issue #3350 for an example of where this happened and was part of
a non-deterministic test failure.

This PR should make (by default) try_assert wait up to 2x as long as
it takes for the assert condition to become true. This is now
parameterised by `factor`, so that it might be set to 1.1 to make
the assert wait for up to 1.1x the time it takes for the assert
condition to become true.