Move more UDP receiver code to own class/file (#3879)
This is part of making radios more pluggable.

After this PR, there is a UDPRadioReceiver, patterned after
ZMQRadioReceiver which contains process state and can be asked to close.

One MonitoringHub shutdown test no longer tests the UDP receiver process
being killed -- the MonitoringHub is no longer directly aware of that
process because of the new UDPRadioReceiver class.

There is a new test that is more focused on the UDP code starting and
exiting correctly, without involving the MonitoringHub, and which runs
much faster.

## Type of change

- Code maintenance/cleanup