Update htex/monitoring fuzz test after PR #3813 (#3886)
PR #3813 made it clearer that monitoring over ZMQ is an HTEX+monitoring
feature, not a general monitoring feature.

This PR renames and rephrases that test in that new context.

PR #3813 also changes that ZMQ channel to listen only on the loopback
address. This PR updates the test to use that address rather than the
previous hub_address attribute.

## Type of change

- Code maintenance/cleanup