Refactor managers to a single ZMQ socket (#3871)
# Description

ZMQ ROUTER and DEALER sockets are two way pipes. I'm not sure why Parsl
has largely ignored that. This commit removes one of the ports.

The two main pieces to address for this change are `interchange.py` and
`process_worker_pool.py`. In the interchange, it's largely to merge the
result handling into what was the "outgoing" method. To disambiguate
what message is what, use the first message of a multipart send to say
what the message contents are (a simple dictionary).

The change in the manager takes the protocol change into account, but
also has to jump through a quick hoop to move the incoming results to
the proper thread for upstream communication: that's the `inproc` PAIR
socket.

After the above two main changes, the rest is "just cleanup," like:

- removing the two ports (task/result) and replacing with `port` or
`worker_port` (singular)

- modifying slightly the command protocol. BINDS is arbitrary, but is
making way for an upcoming second refactor.

- related mypy and test changes

# Changed Behaviour

No workflow should have to change, but with the reduction of a socket
per node manager, the interchange should be able to support larger
swarms.

# Fixes

Fixes #3022 

## Type of change

- Bug fix
- Code maintenance/cleanup