Fix UDPRadio atexit_timeout and use it in tests
This parameter wasn't wired through, but because there was a default
deeper in the code, nothing detected that.

See #2973 for some context around this deep-default coding style: this
PR removes the duplicated defaults and leaves the default only on
the user facing UDPRadio class.

This PR uses the now-working parameter to make UDPRadioReceiver shutdown
faster in tests: because in these test cases, we know that we have
received all messages that we will process, there is no need to wait
for any further outstanding messages. This saves around 9 seconds
of test runtime - both theoretically (3 tests, each 3 second saving)
and in practice on my laptop.

Now that this is wired through properly, a type mismatch between
the original atexit_timeout (int) and the unwired UDPRadio parameter
(float | int) is revealed and fixed: everything uses the original
int type now.