Add HMAC to UDP monitoring messages (#3900)
This is intended to stop spoofing and other misdelivery of UDP messages.

Each UDP receiver generates a new HMAC key, which is then conveyed to
the relevant senders using the config mechanism introduced in PR #3892.
That makes key lifetime roughly as long as the DFK/executor lifetime -
or around the length of the workflow run.

This PR adds two tests around failing validation:
 i) when an incorrect HMAC key is used
 ii) when the message is so short it cannot be parsed into an HMAC+data

## Type of change

- New feature