Detach log handler at shutdown of DFK
Prior to this PR, the automatic parsl.log handlers for each DFK will
accumulate on the "parsl" log topic as new DFKs in one process are
created, so for example the first created parsl.log will contain the
log messages for every parsl workflow run in that process.

This is especially noticeable in pytest tests, but real world users
also occasionally want to run multiple DFKs in one process.

For multiple DFKs in a process that exist completely serially, with
one being shut down completely before the next starts, this should
separate the logs from the two DFKs (and components) entirely.

For DFKs that exist concurrently, this PR does not stop overlapping
logs. The current approach of per-module loggers (rather than
per-DFK/per-object) makes that hard and this PR does not attempt to
solve that.